+++ src/Diagnostics/Diagnostics.jl
@@ -14,7 +14,7 @@ using ..Filaments:
     knots, segments, integrate,
     number_type
 
-using ..BiotSavart: BiotSavart, BiotSavartCache, LongRangeCache, Infinity, ∞,
+using ..BiotSavart: BiotSavart, BiotSavartCache, LongRangeCache, ParamsBiotSavart, Infinity, ∞,
                     ka_generate_kernel, ka_default_workgroupsize
 
 using Bumper: Bumper, @no_escape, @alloc